Removes logs and add copyright information.
