--- Source/EmacsKeyHelper.swift
@@ -1,43 +1,0 @@
-// Copyright (c) 2022 and onwards The McBopomofo Authors.
-//
-// Permission is hereby granted, free of charge, to any person
-// obtaining a copy of this software and associated documentation
-// files (the "Software"), to deal in the Software without
-// restriction, including without limitation the rights to use,
-// copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the
-// Software is furnished to do so, subject to the following
-// conditions:
-//
-// The above copyright notice and this permission notice shall be
-// included in all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-// OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
-// HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-// WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
-// OTHER DEALINGS IN THE SOFTWARE.
-
-import Cocoa
-
-@objc enum McBopomofoEmacsKey: UInt16 {
-    case none = 0
-    case forward = 6 // F
-    case backward = 2 // B
-    case home = 1 // A
-    case end = 5 // E
-    case delete = 4 // D
-    case nextPage = 22 // V
-}
-
-class EmacsKeyHelper: NSObject {
-    @objc static func detect(charCode: UniChar, flags: NSEvent.ModifierFlags) -> McBopomofoEmacsKey {
-        if flags.contains(.control) {
-            return McBopomofoEmacsKey(rawValue: charCode) ?? .none
-        }
-        return .none;
-    }
-}
