+++ Source/KeyHandlerInput.swift
@@ -1,6 +1,29 @@
+// Copyright (c) 2022 and onwards The McBopomofo Authors.
+//
+// Permission is hereby granted, free of charge, to any person
+// obtaining a copy of this software and associated documentation
+// files (the "Software"), to deal in the Software without
+// restriction, including without limitation the rights to use,
+// copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the
+// Software is furnished to do so, subject to the following
+// conditions:
+//
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+// OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+// HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+// WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+// OTHER DEALINGS IN THE SOFTWARE.
+
 import Cocoa
 
-@objc enum KeyCode: Int {
+enum KeyCode: UInt16 {
     case none = 0
     case enter = 76
     case up = 126
@@ -18,18 +41,18 @@ import Cocoa
     @objc private (set) var useVerticalMode: Bool
     @objc private (set) var inputText: String?
     @objc private (set) var charCode: UInt16
-    @objc private var keyCode: Int
-    @objc private var flags: NSEvent.ModifierFlags
-    @objc private var cursorForwardKey: KeyCode
-    @objc private var cursorBackwardKey: KeyCode
-    @objc private var extraChooseCandidateKey: KeyCode
-    @objc private var absorbedArrowKey: KeyCode
-    @objc private var verticalModeOnlyChooseCandidateKey: KeyCode
+    private var keyCode: UInt16
+    private var flags: NSEvent.ModifierFlags
+    private var cursorForwardKey: KeyCode
+    private var cursorBackwardKey: KeyCode
+    private var extraChooseCandidateKey: KeyCode
+    private var absorbedArrowKey: KeyCode
+    private var verticalModeOnlyChooseCandidateKey: KeyCode
     @objc private (set) var emacsKey: McBopomofoEmacsKey
 
     @objc init(inputText: String?, keyCode: UInt16, charCode: UInt16, flags: NSEvent.ModifierFlags, isVerticalMode: Bool) {
         self.inputText = inputText
-        self.keyCode = Int(keyCode)
+        self.keyCode = keyCode
         self.charCode = charCode
         self.flags = flags
         self.useVerticalMode = isVerticalMode
@@ -44,7 +67,7 @@ import Cocoa
 
     @objc init(event: NSEvent, isVerticalMode: Bool) {
         self.inputText = event.characters
-        self.keyCode = Int(event.keyCode)
+        self.keyCode = event.keyCode
         self.flags = event.modifierFlags
         self.useVerticalMode = isVerticalMode
         let charCode: UInt16 = {
@@ -62,6 +85,7 @@ import Cocoa
         self.absorbedArrowKey = useVerticalMode ? .right : .up
         self.verticalModeOnlyChooseCandidateKey = useVerticalMode ? absorbedArrowKey : .none
         super.init()
+        NSLog("self.keyCode \(self.keyCode)")
     }
 
     @objc var isShiftHold: Bool {
@@ -150,3 +174,21 @@ import Cocoa
 
 }
 
+@objc enum McBopomofoEmacsKey: UInt16 {
+    case none = 0
+    case forward = 6 // F
+    case backward = 2 // B
+    case home = 1 // A
+    case end = 5 // E
+    case delete = 4 // D
+    case nextPage = 22 // V
+}
+
+class EmacsKeyHelper: NSObject {
+    @objc static func detect(charCode: UniChar, flags: NSEvent.ModifierFlags) -> McBopomofoEmacsKey {
+        if flags.contains(.control) {
+            return McBopomofoEmacsKey(rawValue: charCode) ?? .none
+        }
+        return .none;
+    }
+}
